[FIX] fixing connexion lost
The driver has been switched to statefull mode. It now keeps the
connexion to the cashology connector at the driver level. It means that
when the posbox restart for any reason the connexion is lost. So adding
the socket None if the config setting are available as a case when the
connexion could reestablished is resolving the issue.